requirements.txt
librosa
soundfile
scikit-learn>=1.3.2
funasr>=1.0.25
moviepy
numpy
gradio
modelscope
torch>=1.13
torchaudio
openai
g4f
dashscope
anthropic
google-generativeai
requests